+++ packages/common/store/actions/index.js
@@ -3,7 +3,7 @@ export * from '@block/store/action';
 export * from '@bookmark/store/action';
 export * from '@network/store/action';
 export * from '@settings/store/actions';
-export * from '@transaction/store/action';
+export * from '@transaction/store/actions';
 export * from '@dpos/store/actions/voting';
 export * from '@dpos/store/actions/watchList';
 export * from '@updater/store/actions/appUpdates';
